requirements.txt
httpx[http2]==0.27.0
selectolax==0.3.17
lxml==5.2.2
readability-lxml==0.8.1
extruct==0.17.0
w3lib==2.2.1
pydantic==2.7.4
pydantic-settings==2.3.4
SQLAlchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
typer==0.12.3
uvicorn==0.30.1
fastapi==0.111.0
python-dateutil==2.9.0.post0
tenacity==9.0.0
structlog==24.1.0
apscheduler==3.10.4
pytest==8.3.1
